Share coordinator config
