+++ coordinator/src/lib.rs
@@ -1,3 +1,4 @@
+pub mod config;
 pub mod faucet;
 pub mod macros;
 pub mod shared_state;
